Fix public events API
Python 3.13 plugin_host of ST4201+ does no longer support `ZipLoader`.

This commit therefore refactors public `package_control` module to use only
built-in import mechanisms if possible and fallback to `ZipLoader` only,
if a python 3.3 plugin needs to load `events` module from
Package Control.sublime-package running on foreign python 3.8 or 3.13 host.

With this change behavior is equal on all plugin_hosts. Only `events` module
is exported as this is the only module intended to be public.